==> 01_data_preparation.ipynb <==
import re
import pandas as pd
import funcs 
import importlib
import unicodecsv as csv
from timeit import default_timer as timer
from datetime import datetime, timedelta

importlib.reload(funcs)

## 0. Wczytanie danych zrzuconych z bazy SQL do csv

politicians = pd.read_csv("00_dane/portraits.csv", sep='|')
#speeches = pd.read_csv("00_dane/speech_data.csv", sep='|')

#wczytujemy plik po modyfikacjach w SQL
speeches = pd.read_csv("00_dane/speeches.csv", sep='|')

## 1. Znalezienie imienia i nazwiska przemawiającego

speeches['author'] = speeches['speech_raw'].apply(funcs.name_finder)

Zapisanie rozszerzonych danych do dalszych analiz.

speeches.to_csv('00_dane/speeches.csv', encoding='UTF=8')

## 2. Czyszczenie tekstu przemówień pt.1


### 2.1. Wczytanie danych po czyszczeniu SQL i dodaniu autora

politicians = pd.read_csv("00_dane/portraits.csv", sep='|')
speeches = pd.read_csv("00_dane/speeches_2.csv", sep='|')

#speeches.to_csv('00_dane/speeches_2.csv', encoding='UTF=8')

len(speeches['speech_raw']), sum(1-pd.isnull(speeches['author_final']))


### 2.2. Czyszczenie znaków specjalnych i wycinanie nagłówka.

start = timer()
speeches['speech_raw'] = speeches['speech_raw'].apply(funcs.rem_head)
speeches['speech_raw'] = speeches['speech_raw'].apply(funcs.clear_text)
speeches['speech_title'] = speeches['speech_title'].apply(funcs.clear_text)
end = timer()
print(end - start)

Wycinam fragment, który jest tytułem:
1. Dodaję "\" przed nawiasami w tytule.
2. Wycinam tytuł z treści.

start = timer()
speeches['speech_title'] = speeches['speech_title'].apply(lambda t: re.sub('\(','\(',t))
speeches['speech_title'] = speeches['speech_title'].apply(lambda t: re.sub('\)','\)',t))
print(timedelta(seconds=timer()-start))

Jeden "problematyczny" case.

speeches = speeches[~speeches['speech_raw'].isnull()]
speeches = speeches[~speeches['speech_title'].isnull()]

speeches['speech_title'] = [re.sub('\[COM\\\\\(2005\\\\\)0119 - C6-0099/2005 - 2005/0043\\\\\(COD\\\\\)\]','',t) for t in speeches['speech_title']]
speeches['speech_raw'] = speeches['speech_raw'].apply(funcs.rem_nasty) 

speeches['speech_raw'] = [funcs.rem_title(t, r) for t, r, in zip(speeches['speech_title'], speeches['speech_raw'])]


### 2.3. Zapisanie zmodyfikowanego pliku

start = timer()
speeches.to_csv('00_dane/speeches_2.csv', encoding='UTF-8', index=False, sep="|")
print(timedelta(seconds=timer()-start))

## 3. Łączenie fragmentów przemówień

%%time
sp_joined = speeches.groupby(['date_', 'author_final','speech_title']).agg({'speech_raw': lambda x: ' '.join(x)}).reset_index()

## 4. Czyszczenie tekstu przemówienia pt 3.


### 4.1. Usuwanie bardziej ogólnych elementów
- Wypowiedzi innych posłów i komentarze typu "Oklaski" pojawiające się w nawiasach
- "Poseł Imię Nazwisko:"

start = timer()
sp_joined['speech_raw'] = sp_joined['speech_raw'].apply(lambda t: re.sub('\([\w.,: ]+\)','',t))
sp_joined['speech_raw'] = sp_joined['speech_raw'].apply(lambda t: re.sub('[Pp]oseł ([\w -\.]+):','',t))
sp_joined['speech_raw'] = sp_joined['speech_raw'].apply(lambda t: re.sub('\s+',' ',t))
print(timedelta(seconds=timer()-start))

### 4.2. Usuwanie przeszkadzających elementów -tworze drugą kolumnę
- Panie Marszałku!
- Pani Marszałek!
- Wysoka Izbo!
- Panie Ministrze!
- Dziękuję bardzo.
- Wysoki Sejmie!
- Panie i Panowie Posłowie!
- Panie Posłanki! Panowie Posłowie!
- Wysokie Zgromadzenie Narodowe!

sp_joined['speech_rawer'] = sp_joined['speech_raw'].apply(lambda t: re.sub('Panie Marszałku!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Szanowny Panie Ministrze!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Pani Marszałek!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Wysoka Izbo!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Panie Ministrze!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Dziękuję bardzo.','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Wysokie Zgromadzenie Narodowe!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Wysoki Sejmie!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Panie Posłanki! Panowie Posłowie!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Panie i Panowie Posłowie!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('\s+',' ',t))

sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Dziękuję[.,] panie marszałku[.,]','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Dziękuję.','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Panie!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('Panowie!','',t))
sp_joined['speech_rawer'] = sp_joined['speech_rawer'].apply(lambda t: re.sub('\s+',' ',t))

Wgranie do pliku

sp_joined.to_csv('00_dane/sp_joined.csv', encoding='UTF-8', index=False, sep="|")
