==> 02_generator_1.0.ipynb <==
# Najprostszy z możliwych model n-gramowy

import re
import pandas as pd
import funcs 
import importlib
import unicodecsv as csv
import pickle
import random
from timeit import default_timer as timer
from datetime import datetime, timedelta
import ng_models
from collections import Counter

importlib.reload(funcs)
importlib.reload(ng_models)

sp_joined = pd.read_csv('00_dane/sp_joined.csv',  sep='|', encoding='UTF-8')

## Spróbujmy ulepszyć model lepiej uwzględniając początek i koniec zdania.

Wszystkie funkcje i metody przeniesione do odrębnego modułu `ng_models`, łącznie z preprocessingiem danych.

dane_do_modelu = ' '.join(sp_joined['speech_raw'])

W pierwszej kolejności model 1-gramowy.

start = timer()
model_1g = ng_models.ng_models(1, dane_do_modelu)
end = timer()
print(timedelta(seconds=timer()-start))

%%time
model_1g.generate(steps=10, max_sent=10)

%%time
model_1g.generate(steps=20, max_sent=10)

del model_1g

Teraz model 2 gramowy

del model_2g

start = timer()
model_2g = ng_models.ng_models(2, dane_do_modelu)
end = timer()
print(timedelta(seconds=timer()-start))

%%time
model_2g.generate(steps=10, max_sent=10)

model_2g.model_dict['są prawnie']

%%time
model_2g.generate(steps=10, max_sent=20)

### Czy uda się model dla n=3?

start = timer()
model_3g = ng_models.ng_models(3, dane_do_modelu)
end = timer()
print(timedelta(seconds=timer()-start))

%%time
model_3g.generate(steps=10, max_sent=20)

a=model_3g.model_dict['mniemanie o politykach']
Counter(a)

### Testowanie perplexity

model_1g.perplexity('Dzisiejsze posiedzenie ma na celu omówienie projektu ustawy')

a=2.9111908274436467e-30

a**(-1/8)

model_1g.perplexity('Lwy to zwierzęta kotowate żyjące na sawannie')

re.findall(r'<s> <s> ([^ ]+)',t)

model_2.V

a = ['<s>', '<s>', 'this']

a[0]

a[0:2]
