==> 03_wstepne_analizy.ipynb <==
import re
import pandas as pd
import numpy as np
import funcs 
import importlib
import unicodecsv as csv
import pickle
import random
from timeit import default_timer as timer
from datetime import datetime, timedelta
import spacy
from spacy.lang.pl import Polish
import nltk
from nltk.tokenize import word_tokenize
import matplotlib.pyplot as plt

from collections import Counter

importlib.reload(funcs)

sp_joined = pd.read_csv('00_dane/sp_joined.csv',  sep='|', encoding='UTF-8')

len(sp_joined)

#### Usuwam cyfry i znaki interpunkcyjne

%%time
sp_test = sp_joined['speech_rawer']
sp_test = sp_test.apply(lambda x: re.sub('[\W\d]',' ', x.lower()))

#czas dla apply : 1 min 10 s

#### Lista stopwords polskich do usunięcia

nlp = spacy.load('pl_model')
stop = set(nlp.Defaults.stop_words)
stop.add('zł')
stop.add('pkt')
stop.add('art')
stop.add('ustawy')
stop.add('r')

#stop = nlp.Defaults.stop_words

%%time
sp_test = sp_test.apply(lambda x: ' '.join([word for word in x.split() if word not in stop]))
#sp_test = [w for w in sp_test if w not in stop]

#Czas dla całości
# 20,9 s


### Ile mamy słów

%%time
corp_dict={}
n = 1
for speech in sp_test:
    if n % 10000 == 0:
        print('Wystąpienie nr: ', n)
    for word in speech.split():
        if word not in corp_dict.keys(): 
            corp_dict[word]=1
        else:
            corp_dict[word]+=1
    n+=1

sp_test.to_csv('00_dane/clear_sp.csv', encoding='UTF=8', sep='|')

len(corp_dict.keys()), sum(corp_dict.values())

#### Sortowanie słownika po liczbie wystąpień

corp_dict = {k: v for k, v in sorted(corp_dict.items(), key=lambda item: item[1])}

corp_dict['suweren']

#### Jaki jest rozkład częstości słów?

Jakie słowa pojawiają się najrzadziej?

u = [k for k,v in corp_dict.items() if v<=5]

len(u)

for _ in range(10):
    print(u[random.randrange(len(u))])

Jakie słowa pojawiają się najczęściej?

sorted(corp_dict.items(), key=lambda item: item[1])[-10:]

#### Ile słów jest w każdym wystąpieniu?

sp_joined['liczba_sl_raw'] = [len(sp.split()) for sp in sp_joined['speech_raw']]

sp_joined['liczba_sl_rawer'] = [len(sp.split()) for sp in sp_joined['speech_rawer']]

sum(sp_joined['liczba_sl_rawer'])

sp_joined.to_csv('00_dane/sp_joined.csv', sep='|', encoding='UTF-8')
