==> 04_generator_2.0_brudnopis.ipynb <==
# Wykorzystanie pakietu `nltk`

import re
import pandas as pd
import funcs 
import importlib
import unicodecsv as csv
import pickle
import random
from timeit import default_timer as timer
from datetime import datetime, timedelta
import nltk
import nltk.lm
from nltk.util import pad_sequence
from nltk.util import bigrams
from nltk.util import ngrams
from nltk.util import everygrams
from nltk.lm.preprocessing import pad_both_ends
from nltk.lm.preprocessing import flatten
from nltk import word_tokenize, sent_tokenize 
from nltk.lm.preprocessing import padded_everygram_pipeline

importlib.reload(funcs)
importlib.reload(nltk)

print(nltk.__version__)

sp_joined = pd.read_csv('00_dane/sp_joined.csv',  sep='|', encoding='UTF-8')

Kilka przykładów zastosowania funkcji generującej n-gramy.

test = [word_tokenize(sent_tokenize(sp_joined['speech_rawer'][0])[0])]

list(bigrams(test[0]))

training_ngrams, padded_sentences = padded_everygram_pipeline(2, test)
for ngramlize_sent in training_ngrams:
    print(list(ngramlize_sent))
    print()
print('#############')
list(padded_sentences)

Przetesumy model na jednej wypowiedzi

text = ''.join(sp_joined['speech_rawer'][:10000])


%%time
tok_text = [list(map(str.lower, word_tokenize(sent))) 
                  for sent in sent_tokenize(text)]

# 10 000 Wall time: 1min 31s

tok_text

train_data, padded_sents = padded_everygram_pipeline(2, tok_text)

padded_sents

from nltk.lm import MLE
model = MLE(2)

%%time
model.fit(train_data, padded_sents)

#10 000 Wall time: 3min 51s

print(model.counts)

model.counts['<\s>']

model.counts[['<\s>']]['<\s>']

model.counts[['moim','zdaniem']]['podjęcie']

print(model.generate(30, text_seed=['<s>']))

model.score('gwarantować', 'zabezpieczenia mają'.split())
