==> 05_modele_per_mowca.ipynb <==
import re
import pandas as pd
import funcs 
import importlib
import unicodecsv as csv
import pickle
import random
from timeit import default_timer as timer
from datetime import datetime, timedelta
import ng_models
import matplotlib.pyplot as plt


%matplotlib notebook
importlib.reload(funcs)
importlib.reload(ng_models)

sp_joined = pd.read_csv('00_dane/sp_joined.csv',  sep='|', encoding='UTF-8')

sp_joined.head(10)

#### Którzy politycy mają nawięcej treści?

ag=sp_joined.groupby('author_final').agg({'liczba_sl_raw': sum, 'date_': 'count'}).reset_index()

sp_joined[sp_joined.author_final=='Zygmunt Kufel']

ag.sort_values('date_', ascending=False).head(5)

ag.sort_values('liczba_sl_raw', ascending=False).head(5)

ag[ag.author_final=='Krystyna Pawłowicz']

ag[ag.author_final=='Donald Tusk']

ag[ag.author_final=='Janusz Korwin-Mikke']

## Budowa modeli dedykowanych poszczególnym posłom

Czy wypowiedzi są podobne?

JK = ' '.join(sp_joined['speech_raw'][sp_joined.author_final=='Jan Kulas'])
JZ = ' '.join(sp_joined['speech_raw'][sp_joined.author_final=='Józef Zych'])
JJ = ' '.join(sp_joined['speech_raw'][sp_joined.author_final=='Jerzy Jaskiernia'])
SS = ' '.join(sp_joined['speech_raw'][sp_joined.author_final=='Stanisław Stec'])
TT = ' '.join(sp_joined['speech_raw'][sp_joined.author_final=='Tadeusz Tomaszewski'])
KP = ' '.join(sp_joined['speech_raw'][sp_joined.author_final=='Krystyna Pawłowicz'])
DT = ' '.join(sp_joined['speech_raw'][sp_joined.author_final=='Donald Tusk'])
JKM = ' '.join(sp_joined['speech_raw'][sp_joined.author_final=='Janusz Korwin-Mikke'])

model_JK = ng_models.ng_models(1, JK)
model_JZ = ng_models.ng_models(1, JZ)
model_JJ = ng_models.ng_models(1, JJ)
model_SS = ng_models.ng_models(1, SS)
model_TT = ng_models.ng_models(1, TT)
model_KP = ng_models.ng_models(1, KP)
model_DT = ng_models.ng_models(1, DT)
model_JKM = ng_models.ng_models(1, JKM)

summary_table = pd.DataFrame()
summary_table['Zdanie']=['Zapewniam państwa, że jestem uczciwym politykiem', 
                        'Lew czarownica i stara szafa',
                        'Zupełnie przypadkowy układ pięciu słów',
                        'Panie ministrze proszę o wysłuchanie',
                        'Polskie rolnictwo rozwija się szybko i ma się dobrze',
                        'Demokracja jest najlepszym systemem politycznym',
                        'Jestem najlepszym kandydatem na prezydenta',
                        'Dla porównania zdanie długie o polityce o ustawie dotyczące zagadnień istotnych, które każdy poseł wypowiedzieć może',
                        'wysoka izbo']

summary_table['JK'] = summary_table['Zdanie'].apply(model_JK.perplexity).round()
summary_table['JZ'] = summary_table['Zdanie'].apply(model_JZ.perplexity).round()
summary_table['JJ'] = summary_table['Zdanie'].apply(model_JJ.perplexity).round()
summary_table['SS'] = summary_table['Zdanie'].apply(model_SS.perplexity).round()
summary_table['TT'] = summary_table['Zdanie'].apply(model_TT.perplexity).round()
summary_table['KP'] = summary_table['Zdanie'].apply(model_KP.perplexity).round()
summary_table['DT'] = summary_table['Zdanie'].apply(model_DT.perplexity).round()
summary_table['JKM'] = summary_table['Zdanie'].apply(model_JKM.perplexity).round()

summary_table.set_index('Zdanie').sort_values('JK')

model_JJ.model_dict['demokracja']

## Najbardziej prawdopodobne zdanie dla danej osoby

Wykorzystuję bardziej oczyszczony tekst aby nie wpadać ciagne na zbitki typu "Wysoka Izbo!".

JK = ' '.join(sp_joined['speech_rawer'][sp_joined.author_final=='Jan Kulas'])
JZ = ' '.join(sp_joined['speech_rawer'][sp_joined.author_final=='Józef Zych'])
JJ = ' '.join(sp_joined['speech_rawer'][sp_joined.author_final=='Jerzy Jaskiernia'])
SS = ' '.join(sp_joined['speech_rawer'][sp_joined.author_final=='Stanisław Stec'])
TT = ' '.join(sp_joined['speech_rawer'][sp_joined.author_final=='Tadeusz Tomaszewski'])
KP = ' '.join(sp_joined['speech_rawer'][sp_joined.author_final=='Krystyna Pawłowicz'])
DT = ' '.join(sp_joined['speech_rawer'][sp_joined.author_final=='Donald Tusk'])
JKM = ' '.join(sp_joined['speech_rawer'][sp_joined.author_final=='Janusz Korwin-Mikke'])
model_JK = ng_models.ng_models(1, JK)
model_JZ = ng_models.ng_models(1, JZ)
model_JJ = ng_models.ng_models(1, JJ)
model_SS = ng_models.ng_models(1, SS)
model_TT = ng_models.ng_models(1, TT)
model_KP = ng_models.ng_models(1, KP)
model_DT = ng_models.ng_models(1, DT)
model_JKM = ng_models.ng_models(1, JKM)

max_sent = pd.DataFrame()
max_sent['start']=['Ja', 'Dziś', 'Sejm','']
max_sent['JK'] = max_sent['start'].apply(model_JK.mprob_sent)
max_sent['JZ'] = max_sent['start'].apply(model_JZ.mprob_sent)
max_sent['JJ'] = max_sent['start'].apply(model_JJ.mprob_sent)
max_sent['SS'] = max_sent['start'].apply(model_SS.mprob_sent)
max_sent['TT'] = max_sent['start'].apply(model_TT.mprob_sent)
max_sent['KP'] = max_sent['start'].apply(model_KP.mprob_sent)
max_sent['DT'] = max_sent['start'].apply(model_DT.mprob_sent)
max_sent['JKM'] = max_sent['start'].apply(model_JKM.mprob_sent)
max_sent.set_index('start')

pd.options.display.max_colwidth=500
