==> 05_modele_per_partia.ipynb <==
import re
import pandas as pd
import funcs 
import importlib
import unicodecsv as csv
import pickle
import random
from timeit import default_timer as timer
from datetime import datetime, timedelta
import ng_models
import matplotlib.pyplot as plt
from collections import Counter

%matplotlib notebook
importlib.reload(funcs)
importlib.reload(ng_models)

len(sp_joined)


sp_joined = pd.read_csv('00_dane/sp_joined2.csv',  sep='|', encoding='UTF-8')


Wgranie pliku przygotowanego w SQL, zawierającego najnowsze przypisanie do partii.

parties = pd.read_csv('00_dane/party.csv',  sep='|', encoding='UTF-8')

parties.head(5)

sp_joined=sp_joined.merge(parties,how='left',left_on=['author_final'],right_on=['full_name'])

sp_joined.head(20)

Zapisanie pliku z dodaną partią.

sp_joined.to_csv('00_dane/sp_joined2.csv', encoding='UTF-8', index=False, sep="|")

#### Które partie mają nawięcej treści?

ag=sp_joined.groupby('party_final').agg({'liczba_sl_raw': sum, 'date_': 'count'}).reset_index()


ag.sort_values('date_', ascending=False).head(7)

## Budowa modeli dedykowanych poszczególnym partiom

Czy wypowiedzi są podobne?

PiS = ' '.join(sp_joined['speech_raw'][sp_joined.party_final=='Prawo i Sprawiedliwość'])
PO = ' '.join(sp_joined['speech_raw'][sp_joined.party_final=='Platforma Obywatelska'])
SLD= ' '.join(sp_joined['speech_raw'][sp_joined.party_final=='Sojusz Lewicy Demokratycznej'])
PSL = ' '.join(sp_joined['speech_raw'][sp_joined.party_final=='Polskie Stronnictwo Ludowe'])
PNZ = ' '.join(sp_joined['speech_raw'][sp_joined.party_final=='posłowie niezrzeszeni'])
AWS = ' '.join(sp_joined['speech_raw'][sp_joined.party_final=='Akcja Wyborcza Solidarność'])
SO = ' '.join(sp_joined['speech_raw'][sp_joined.party_final=='Samoobrona'])


### Czy modele wygenerują się wystarczająco szybko?

%%time
model_PiS = ng_models.ng_models(1, PiS)
model_PO = ng_models.ng_models(1, PO)
model_SLD = ng_models.ng_models(1, SLD)
model_PSL = ng_models.ng_models(1, PSL)
model_PNZ = ng_models.ng_models(1, PNZ)
model_AWS = ng_models.ng_models(1, AWS)
model_SO = ng_models.ng_models(1, SO)

%%time
model_PiS = ng_models.ng_models(2, PiS)
model_PO = ng_models.ng_models(2, PO)
model_SLD = ng_models.ng_models(2, SLD)
model_PSL = ng_models.ng_models(2, PSL)
model_PNZ = ng_models.ng_models(2, PNZ)
model_AWS = ng_models.ng_models(2, AWS)
model_SO = ng_models.ng_models(2, SO)

#dla 1-gram 2 min 44 sek 9na nowym kompie 1 min 15 s
# dla 2 gram Wall time: 2min 22s

%%time
model_PO = ng_models.ng_models(3, PO)
model_SO = ng_models.ng_models(3, SO)

#dla 1-gram 2 min 44 sek 9na nowym kompie 1 min 15 s
# dla 2 gram Wall time: 2min 22s

summary_table = pd.DataFrame()
summary_table['Zdanie']=['Zapewniam państwa, że jestem uczciwym politykiem', 
                        'Lew czarownica i stara szafa',
                        'Zupełnie przypadkowy układ pięciu słów',
                        'Panie ministrze proszę o wysłuchanie',
                        'Polskie rolnictwo rozwija się szybko i ma się dobrze',
                        'Demokracja jest najlepszym systemem politycznym',
                        'Jestem najlepszym kandydatem na prezydenta',
                        'Dla porównania zdanie długie o polityce o ustawie dotyczące zagadnień istotnych, które każdy poseł wypowiedzieć może',
                        'wysoka izbo']

summary_table['PO'] = summary_table['Zdanie'].apply(model_PO.perplexity).round()

summary_table['SO'] = summary_table['Zdanie'].apply(model_SO.perplexity).round()

summary_table['PiS'] = summary_table['Zdanie'].apply(model_PiS.perplexity).round()
summary_table['PO'] = summary_table['Zdanie'].apply(model_PO.perplexity).round()
summary_table['SLD'] = summary_table['Zdanie'].apply(model_SLD.perplexity).round()
summary_table['PSL'] = summary_table['Zdanie'].apply(model_PSL.perplexity).round()
summary_table['PNZ'] = summary_table['Zdanie'].apply(model_PNZ.perplexity).round()
summary_table['AWS'] = summary_table['Zdanie'].apply(model_AWS.perplexity).round()
summary_table['SO'] = summary_table['Zdanie'].apply(model_SO.perplexity).round()

summary_table.set_index('Zdanie').sort_values('PiS')

model_PO.model_dict['test z']

model_PO.generate(steps=10, max_sent=20)

model_SO.generate(steps=10, max_sent=10)

model_PO.mprob_sent('')

model_SO.mprob_sent('start')

## Najbardziej prawdopodobne zdanie dla danej partii

Wykorzystuję bardziej oczyszczony tekst aby nie wpadać ciagne na zbitki typu "Wysoka Izbo!".

PiS = ' '.join(sp_joined['speech_rawer'][sp_joined.party_final=='Prawo i Sprawiedliwość'])
PO = ' '.join(sp_joined['speech_rawer'][sp_joined.party_final=='Platforma Obywatelska'])
SLD= ' '.join(sp_joined['speech_rawer'][sp_joined.party_final=='Sojusz Lewicy Demokratycznej'])
PSL = ' '.join(sp_joined['speech_rawer'][sp_joined.party_final=='Polskie Stronnictwo Ludowe'])
PNZ = ' '.join(sp_joined['speech_rawer'][sp_joined.party_final=='posłowie niezrzeszeni'])
AWS = ' '.join(sp_joined['speech_rawer'][sp_joined.party_final=='Akcja Wyborcza Solidarność'])
SO = ' '.join(sp_joined['speech_rawer'][sp_joined.party_final=='Samoobrona'])


PiS[:1000]

%%time
model_PiS = ng_models.ng_models(1, PiS)
model_PO = ng_models.ng_models(1, PO)
model_SLD = ng_models.ng_models(1, SLD)
model_PSL = ng_models.ng_models(1, PSL)
model_PNZ = ng_models.ng_models(1, PNZ)
model_AWS = ng_models.ng_models(1, AWS)
model_SO = ng_models.ng_models(1, SO)


%%time
model_PiS = ng_models.ng_models(2, PiS)
model_PO = ng_models.ng_models(2, PO)
model_SLD = ng_models.ng_models(2, SLD)
model_PSL = ng_models.ng_models(2, PSL)
model_PNZ = ng_models.ng_models(2, PNZ)
model_AWS = ng_models.ng_models(2, AWS)
model_SO = ng_models.ng_models(2, SO)

max_sent = pd.DataFrame()
max_sent['start']=['Ja nie', 'Dziś jest', 'Czy już','Nie można', '']
max_sent['PiS'] = max_sent['start'].apply(model_PiS.mprob_sent)
max_sent['PO'] = max_sent['start'].apply(model_PO.mprob_sent)
max_sent['SLD'] = max_sent['start'].apply(model_SLD.mprob_sent)
max_sent['PSL'] = max_sent['start'].apply(model_PSL.mprob_sent)
max_sent['PNZ'] = max_sent['start'].apply(model_PNZ.mprob_sent)
max_sent['AWS'] = max_sent['start'].apply(model_AWS.mprob_sent)
max_sent['SO'] = max_sent['start'].apply(model_SO.mprob_sent)

max_sent.set_index('start')

pd.options.display.max_colwidth=500


Counter(model_PiS.model_dict['ja nie']).most_common(1)

Counter(model_PO.model_dict['ja']).most_common(1)

Counter(model_SO.model_dict['ja']).most_common(1)

PiS = ' '.join(sp_joined['speech_rawer'][sp_joined.party_final=='Prawo i Sprawiedliwość'])
model_PiS = ng_models.ng_models(2, PiS)

model_PiS.mprob_sent('ja nie')
