==> 06_lematyzacja.ipynb <==
import re
import pandas as pd
import funcs 
import importlib
import unicodecsv as csv
import pickle
import random
from timeit import default_timer as timer
from datetime import datetime, timedelta
import ng_models
import matplotlib.pyplot as plt
from collections import Counter
import stanza

#stanza.download('pl')

%matplotlib notebook
importlib.reload(funcs)
importlib.reload(ng_models)

stopwords = open('00_dane/polish.stopwords.txt','r').read().split()
stopwords.append(['.', ',','!', '?'])

s = pd.read_csv('00_dane/sp_joined.csv',  sep='|', encoding='UTF-8')

sp =pd.read_csv('00_dane/rawest.csv', sep='|', encoding='UTF-8')

sp = sp.drop(columns = 'Unnamed: 0', axis=1)
#sp.columns

sp.iloc[66666]['speech_rawest']

sp.tail(10)

sp_joined = pd.read_csv('00_dane/sp_joined2.csv',  sep='|', encoding='UTF-8')

## Przygotowanie zbioru danych pod analizę tematyczną (i poźniej pod analizę sentymentu również):
- lematyzacja
- usunięcie stop words
- wektoryzacja

sp_joined.iloc[6]['speech_raw']

sp = sp_joined[['date_', 'speech_rawer', 'full_name', 'party_final']]

sp.tail(10)

### Usunięcie stop words oraz lematyzacja

nlp = stanza.Pipeline('pl')

%%time
x=0
for i in sp[sp['speech_rawest'].isnull()].index:
    
    doc = nlp(sp.loc[i, 'speech_rawer'])
    sp.loc[i, 'speech_rawest']= ' '.join([word.lemma for sent in doc.sentences for word in sent.words if word.lemma not in stopwords])
    if round(100*i/len(sp['speech_rawer']),0)>x:
        x = round(100*i/len(sp['speech_rawer']),0)
        print('Done: ',x,'%')
        sp.to_csv('00_dane/rawest.csv', sep='|', encoding='UTF-8')

sp.to_csv('00_dane/rawest.csv', sep='|', encoding='UTF-8', index=False)


sp[sp['speech_rawest'].isnull()].index

sp[sp['speech_rawer']==' '].index

nlp("sportowy")

sp = sp.drop(index=80002)
