==> 11_author_topic_models.ipynb <==
import re
import pandas as pd
import numpy as np
import funcs 
import importlib
import unicodecsv as csv
import pickle
import random
import gensim
from gensim import corpora, models
from timeit import default_timer as timer
from datetime import datetime, timedelta
import nltk
from nltk.tokenize import word_tokenize
import matplotlib.pyplot as plt
import spacy
from collections import Counter
import pyLDAvis
import pyLDAvis.gensim
from sklearn.model_selection import train_test_split

importlib.reload(funcs)

dane = pd.read_csv('00_dane/dane_TM.csv',  sep='|', encoding='UTF-8')
dane['speech_rawest'] = dane['speech_rawest'].apply(lambda x: re.sub(r'[\W\d]',' ', x.lower()))
nlp = spacy.load('pl_model')
stop = set(nlp.Defaults.stop_words)
stop.add('zł')
stop.add('pkt')
stop.add('art')
stop.add('ustawy')
stop.add('r')
dane['speech_rawest'] = dane['speech_rawest'].apply(lambda x: [word for word in x.split() if word not in stop])

Ponieważ pusta wartość jest interpretowana jako float, należy ograniczyć zbiór.

dane = dane[dane['full_name'].apply(lambda x:isinstance(x, str))]

dane = dane.reset_index()

dane = dane.drop('index', axis=1)

dane.head(3)

dane[dane['party_final']=='Platforma Obywatelska']

Utworzenie słownika autor - dokument

## Przygotowanie słowników itp.

dane.head()

aut2doc=dict()
for (aut, doc_nr) in zip(dane['full_name'], dane.index):
    if aut not in aut2doc.keys():
        aut2doc[aut]=[]
    aut2doc[aut].append(doc_nr)

## Słowniki i dictionary po utworzeniu są zrzucone do pickli...

aut2doc=dict()
for (aut, doc_nr) in zip(dane['full_name'], dane.index):
    if aut not in aut2doc.keys():
        aut2doc[aut]=[]
    aut2doc[aut].append(doc_nr)
doc2aut = dane['full_name'].T.to_dict()
corp = dane['speech_rawest']
bigram = gensim.models.Phrases(corp, min_count=20, threshold=50)
bigram_mod = gensim.models.phrases.Phraser(bigram)
corp = [bigram_mod[doc] for doc in corp]
dictionary = gensim.corpora.Dictionary(corp)
dictionary.filter_extremes(no_above=0.2, no_below=50)
dictionary.compactify()
pickle.dump(dictionary, open('dict_atm.p', 'wb'))
BOW = [dictionary.doc2bow(text) for text in corp]
pickle.dump(BOW, open('bow_atm.p', 'wb'))

from gensim.corpora import mmcorpus
from gensim.test.utils import common_dictionary, datapath, temporary_file
print(mmcorpus.MmCorpus(datapath('testcorpus.mm'))[8])

from gensim.models import AuthorTopicModel
%time model = AuthorTopicModel(corpus=BOW, num_topics=30, id2word=dictionary.id2token, 
                               author2doc=aut2doc, passes=5, eval_every=0, iterations=1)

pickle.dump(model, open('model_atm.p', 'wb'))

## ..... i ładowane

model = pickle.load(open('model_atm.p', 'rb'))
dictionary = pickle.load(open('dict_atm.p', 'rb'))
BOW = pickle.load(open('bow_atm.p', 'rb'))

## Opis tematów

for i in range(30):
    print('TOPIC ', i)
    for a, b in model.show_topic(i):
        print(a)

topics = ['rynek mieszkaniowy', 'umowy gospodarcze', 'służby państwowe', 'naród - patos', '4', 'świadczenia pomocowe', 
          'europa', 'edukacja', '8', 'gospodarka', 'UE', 'sądownictwo', 'służba zdrowia', 'konstytucja', 'wybory', 'posiedzenie'
         , 'budżet', 'rolnictwo', 'ochrona środowiska', 'wymiar prawa', 'dyskusja', 'polityka', 'finanse', 'przedsiębiorstwa',
         'samorządy', 'miasta', 'drogi', 'ustrój państwa', '28', 'rząd']

Tematy są trudne do opisania. Być może mówców jest zbyt dużo i przez to się rozdrabnia?
Później zweryfikujemy czy lepiej wychodzi dla partii.

pd.DataFrame(model.get_author_topics('Elżbieta Witek'))

pd.DataFrame(topics)

## Model dla partii

aut2doc=dict()
for (aut, doc_nr) in zip(dane['party_final'], dane.index):
    if aut not in aut2doc.keys():
        aut2doc[aut]=[]
    aut2doc[aut].append(doc_nr)
doc2aut = dane['party_final'].T.to_dict()

Model wygenerowany na parametrach takich jak poprzednio był bardzo trudny do opisu.
Spróbuję innych wersji - z mniejsza liczbą tematów, albo większą liczbą iteracji.

dictionary.id2token

from gensim.models import AuthorTopicModel
%time model1 = AuthorTopicModel(corpus=BOW, num_topics=20, id2word=dictionary.id2token, author2doc=aut2doc, passes=3, eval_every=0, iterations=1)

pickle.dump(model1, open('model_ptm_1.p', 'wb'))

dictionary = pickle.load(open('dict_atm.p', 'rb'))
BOW = pickle.load(open('bow_atm.p', 'rb'))
model1 = pickle.load(open('model_ptm_1.p', 'rb'))

for i in range(20):
    print('TOPIC ', i)
    for a, b in model1.show_topic(i):
        print(a)

topics_p = ['narodowy', 'polityka', 'budżet', '3', '4', 'górnictwo', '']
