==> 15_poprawka_przypisania_sent.ipynb <==
# Inne podejście do przypisania sentymentu

* Uwzględnienie aby słowa nie były zbyt blisko innych kategorii.
* Aby odległość nie była zbyt duża
* Dalej pracujemy z wykluczoną kategorią neutralną

import re
import pandas as pd
import numpy as np
import funcs 
import importlib
import unicodecsv as csv
import pickle
import random
import gensim
from gensim import corpora, models
from timeit import default_timer as timer
from datetime import datetime, timedelta
import nltk
from nltk.tokenize import word_tokenize
import matplotlib.pyplot as plt
import spacy
from collections import Counter
import pyLDAvis
import pyLDAvis.gensim
from sklearn.model_selection import train_test_split
import scipy

importlib.reload(funcs)

%time dane = pd.read_csv('00_dane/dane_cleanest.csv',  sep='|', encoding='UTF-8')
%time dane['speech_rawest'] = dane['speech_rawest'].apply(lambda x: re.sub(r'[\W\d]',' ', x))
%time dane['speech_rawest'] = dane['speech_rawest'].apply(lambda x: [word for word in x.split()])

## 1. Wczytanie listy słów z NAWL

Parametry wg których przypisano kategorie zmieniono aby zwiększyć liczbę przypisań. Usunięto grupę 'neutral'.

NAWL = pd.read_csv('00_dane/nawl-analysis_2.csv', encoding='UTF-8')
NAWL.drop('distance to N', axis=1, inplace=True)
NAWL.head()

print('Lista zawiera {} słów.'.format(len(NAWL)))

Analiza wczytanych danych

NAWL.groupby('category')["word"].count()

Klasa "U" oznacza brak klasyfikacji. Mamy więc mało przypisanych słów.

Tu zmiana - analiza jaka powinna być minimalna odległość do innych kategorii.

Inne:
- Happiness
- Anger
- Sadness
- Fear
- Neutral (zmia=eniono parametry tak, aby się nie dopsiywało)
- Disgust


NAWL[NAWL.category=='H'].describe()

NAWL[NAWL.category=='A'].describe()

NAWL[NAWL.category=='S'].describe()

NAWL[NAWL.category=='F'].describe()

NAWL[NAWL.category=='D'].describe()

a = NAWL.iloc[100]
a

Przypiszmy mimo wszystko flagę, do której odległość jest najmniejsza.

Dodajemy dwa dodatkowe parametry, które pozwolą zminimalizować złe przypisania i zachować rozdzielność zbiorów.

NAWL['new_flag']=''
for i in range(len(NAWL)):
    a = NAWL.iloc[i]
    l = [a['distance to H'], a['distance to A'], a['distance to S'], a['distance to F'], a['distance to D']]
    l.sort()
    m=min(l)
    if (l[1]-m>0.5) & (m<6.2):
        if m == a['distance to H']:
            NAWL.loc[i, 'new_flag']='H'
        elif m == a['distance to S']:
            NAWL.loc[i, 'new_flag']='S'
        elif m == a['distance to F']:
            NAWL.loc[i, 'new_flag']='F'
        elif m == a['distance to D']:
            NAWL.loc[i, 'new_flag']='D'
        elif m == a['distance to A']:
            NAWL.loc[i, 'new_flag']='A'

NAWL.groupby('new_flag')["word"].count()

#new_flag
#A     322
#D     149
#F     465
#H    1740
#S     226
#Name: word, dtype: int64'''

NAWL[(NAWL.new_flag=='') & (NAWL.category!='U')]

NAWL = NAWL[NAWL.new_flag!='']

NAWL.shape

## 2. Przypisanie sentymentu słów

### 2.1 Budowa słownika wszystkich słów

dict_sp = {}
for sp in dane['speech_rawest']:
    for w in sp:
        if w not in dict_sp.keys():
            dict_sp[w]=0
        dict_sp[w]+=1

print('Liczba unikalnych słów w słowniku wynosi {}.'.format(len(dict_sp)))

df_sp = pd.DataFrame(list(dict_sp.items()), columns=['word', 'counter'])
df_sp.head()

len(df_sp)

### 2.2 Dogranie parametrów

df_sp = pd.merge(df_sp, NAWL, left_on='word', right_on='word', how='left' )

df_sp[300:310]

Ile mamy dogranych?

len(df_sp[pd.isnull(df_sp['category'])==False])

pickle.dump(df_sp, open('df_sp.p', 'wb'))

## 3. Dogranie embeddingów

df_sp = pickle.load(open('df_sp.p','rb'))

df_sp.head()


### 3.1 Pobranie modelu dla języka polskiego

import fasttext.util
#fasttext.util.download_model('pl', if_exists='ignore')  
ft = fasttext.load_model('cc.pl.300.bin')

### 3.2 Słowniki pomocnicze

NAWL_vec_dict = {}
for w in NAWL["word"]:
    NAWL_vec_dict[w] = ft.get_word_vector(w)
corp_vec_dict = {}
for w in df_sp["word"]:
    corp_vec_dict[w] = ft.get_word_vector(w)    

pickle.dump(NAWL_vec_dict, open('nawl.p', 'wb'))
pickle.dump(corp_vec_dict, open('corp_vec.p', 'wb'))
                        

### 3.3 Przypisanie sentymentu na podstawie pięciu najbliższych słów z listy NAWL

neigh_cnt =5 #parametr z ilu słów korzystamy
max_dist = 0.7

for word in df_sp['word'][pd.isnull(df_sp['new_flag'])==True]:

    word_vec = corp_vec_dict[word]
    word_dict = {}
    neigh_df = pd.DataFrame(columns=['distance to H','distance to A','distance to S','distance to F','distance to D'])
    
    #obliczamy odległość do każdego słowa z NAWL
    for k in NAWL_vec_dict.keys():
        word_dict[k] = scipy.spatial.distance.cosine(word_vec, NAWL_vec_dict[k])

    #dla 'neigh_cnt' najbliższym (pod warunkiem, że odległość nie jest bardzo duża)
    for neighbor, dist in sorted(word_dict.items(), key=lambda item: item[1])[:neigh_cnt]:
        if dist<=max_dist:
            neigh_df = neigh_df.append(NAWL[['distance to H','distance to A','distance to S','distance to F','distance to D']][NAWL['word']==neighbor] )

    n_stats = neigh_df.mean()
    if len(neigh_df>0):
        df_sp.loc[df_sp.word==word, 'distance to H'] = n_stats['distance to H']
        df_sp.loc[df_sp.word==word, 'distance to A'] = n_stats['distance to A']
        df_sp.loc[df_sp.word==word, 'distance to S'] = n_stats['distance to S']
        df_sp.loc[df_sp.word==word, 'distance to F'] = n_stats['distance to F']
        df_sp.loc[df_sp.word==word, 'distance to D'] = n_stats['distance to D']

       
            
        l = list(neigh_df.mean())
        l.sort()
        m=min(l)
        if (l[1]-m>0.5) & (m<6.2):
            if m == neigh_df.mean()['distance to H']:
                df_sp.loc[df_sp.word==word, 'new_flag']='H'
            elif m == neigh_df.mean()['distance to S']:
                df_sp.loc[df_sp.word==word, 'new_flag']='S'
            elif m == neigh_df.mean()['distance to F']:
                df_sp.loc[df_sp.word==word, 'new_flag']='F'
            elif m == neigh_df.mean()['distance to D']:
                df_sp.loc[df_sp.word==word, 'new_flag']='D'
            elif m == neigh_df.mean()['distance to A']:
                df_sp.loc[df_sp.word==word, 'new_flag']='A'

print("Uzupełniony sentyment dla {} z {} słów.".format(len(df_sp[pd.isnull(df_sp['new_flag'])==False]), len(df_sp)))

pickle.dump(df_sp, open('df_sp.p', 'wb'))

df_sp[df_sp.word=='skutek']
